Fix unknown column size, such as for text columns
It now returns Nothing

From: Matthias Radestock
I have a PostgreSQL database with tables containing columns of type
TEXT. When I ask HDBC to describe the table it reports the column size
as (Just -1). The query executed by HDBC to obtain this information is

  select attname, atttypid, attlen, attnum, attnotnull
  from pg_attribute, pg_class
  where relname = 'MyTable' and attnum > 0 and attrelid = pg_class.oid;

This does indeed return -1 for the attlen of a TEXT column.

My question is: Should HDBC return Nothing instead?

I am asking because the above behaviour breaks haskelldb's DBDirect code
generation. It generates references to a type "BStr-1". If the returned
value is Nothing instead then haskelldb recognises the column type as a
string of unknown length and does the right thing.