As an alternative to hard-coding, use pg_config to find the includedir
This means we have to use a custom Setup.hs which is a bit fragile.
On the other hand hard coding the pg include dir is not going to work on
al systems. So it's a trade off, you choose. For example on my system
pg_config says to use /usr/include/postgresql/libpq-4 as the inc dir.