Relax the system_schema exclusion
From: MightyByte (Doug Beardsley)
The database I was using for
reference had system_schema tables that made me think they were
Postgres-specific, but really were a part of the database.  So the
last AND clause for "system_schema" should be removed.  It looks like
the "pg_catalog" and "information_schema" schemas really are for
internal Postgres stuff, but you might want to double check me to make
sure.