Mark the FFI calls pqconnectdb, pqexecParams and pqexec as 'safe' to allow concurrent access to the database.
Previously all FFI calls were marked 'unsafe' which meant any long running db
query would "stop the world".  That is, the whole haskell runtime system is
blocked until the database query completes.  This is undesirable in many
multi-threaded programs, and especially so in "server" programs where it is
also necessary to be able to have multiple active database queries.

- Added an MVar to the Conn type.  This is used to ensure each Connection is
  only used with one FFI at a time.  Multiple database queries may be performed
  using separate Connections.

- In the case where the local libpq is not compiled to be thread-safe, issue a
  warning to stderr, and use a global MVar to allow only a single FFI call to
  be active at once.