This adds support for escaping the question mark in queries.
This is necessary to write queries of the form:

    SELECT id FROM nodes WHERE tags ? 'name';

Here the table 'nodes' has a collumn of type HStore named 'tags'.
The question mark is an opperator that check if the key 'name' exists in
the HStore collumn.

To write the above query in HDBC:

    quickQuery' conn "SELECT id FROM nodes WHERE tags \\? 'name'" []

For more information about HStore collumns, check section F.13.2 on
http://www.postgresql.org/docs/8.4/static/hstore.html